+++ crates/miden-protocol/src/batch/proven_batch.rs
@@ -7,12 +7,7 @@ use crate::batch::{BatchAccountUpdate, BatchId};
 use crate::block::BlockNumber;
 use crate::errors::ProvenBatchError;
 use crate::note::Nullifier;
-use crate::transaction::{
-    InputNoteCommitment,
-    InputNotes,
-    OrderedTransactionHeaders,
-    ProvenOutputNote,
-};
+use crate::transaction::{InputNoteCommitment, InputNotes, OrderedTransactionHeaders, OutputNote};
 use crate::utils::serde::{
     ByteReader,
     ByteWriter,
@@ -32,7 +27,7 @@ pub struct ProvenBatch {
     reference_block_num: BlockNumber,
     account_updates: BTreeMap<AccountId, BatchAccountUpdate>,
     input_notes: InputNotes<InputNoteCommitment>,
-    output_notes: Vec<ProvenOutputNote>,
+    output_notes: Vec<OutputNote>,
     batch_expiration_block_num: BlockNumber,
     transactions: OrderedTransactionHeaders,
 }
@@ -53,7 +48,7 @@ impl ProvenBatch {
         reference_block_num: BlockNumber,
         account_updates: BTreeMap<AccountId, BatchAccountUpdate>,
         input_notes: InputNotes<InputNoteCommitment>,
-        output_notes: Vec<ProvenOutputNote>,
+        output_notes: Vec<OutputNote>,
         batch_expiration_block_num: BlockNumber,
         transactions: OrderedTransactionHeaders,
     ) -> Result<Self, ProvenBatchError> {
@@ -137,7 +132,7 @@ impl ProvenBatch {
     ///
     /// This is the aggregation of all output notes by the transactions in the batch, except the
     /// ones that were consumed within the batch itself.
-    pub fn output_notes(&self) -> &[ProvenOutputNote] {
+    pub fn output_notes(&self) -> &[OutputNote] {
         &self.output_notes
     }
 
@@ -178,7 +173,7 @@ impl Deserializable for ProvenBatch {
         let reference_block_num = BlockNumber::read_from(source)?;
         let account_updates = BTreeMap::read_from(source)?;
         let input_notes = InputNotes::<InputNoteCommitment>::read_from(source)?;
-        let output_notes = Vec::<ProvenOutputNote>::read_from(source)?;
+        let output_notes = Vec::<OutputNote>::read_from(source)?;
         let batch_expiration_block_num = BlockNumber::read_from(source)?;
         let transactions = OrderedTransactionHeaders::read_from(source)?;