+++ crates/miden-testing/tests/scripts/send_note.rs
@@ -16,7 +16,7 @@ use miden_protocol::note::{
     NoteType,
     PartialNote,
 };
-use miden_protocol::transaction::OutputNote;
+use miden_protocol::transaction::RawOutputNote;
 use miden_protocol::{Felt, Word};
 use miden_standards::account::interface::{AccountInterface, AccountInterfaceExt};
 use miden_standards::code_builder::CodeBuilder;
@@ -82,7 +82,7 @@ async fn test_send_note_script_basic_wallet() -> anyhow::Result<()> {
         .build_tx_context(sender_basic_wallet_account.id(), &[spawn_note.id()], &[])
         .expect("failed to build tx context")
         .tx_script(send_note_transaction_script)
-        .extend_expected_output_notes(vec![OutputNote::Full(note.clone())])
+        .extend_expected_output_notes(vec![RawOutputNote::Full(note.clone())])
         .build()?
         .execute()
         .await?;
@@ -107,9 +107,9 @@ async fn test_send_note_script_basic_wallet() -> anyhow::Result<()> {
     );
     assert_eq!(
         executed_transaction.output_notes().get_note(0),
-        &OutputNote::Partial(p2any_note.into())
+        &RawOutputNote::Partial(p2any_note.into())
     );
-    assert_eq!(executed_transaction.output_notes().get_note(1), &OutputNote::Full(note));
+    assert_eq!(executed_transaction.output_notes().get_note(1), &RawOutputNote::Full(note));
 
     Ok(())
 }
@@ -157,12 +157,12 @@ async fn test_send_note_script_basic_fungible_faucet() -> anyhow::Result<()> {
         .build_tx_context(sender_basic_fungible_faucet_account.id(), &[], &[])
         .expect("failed to build tx context")
         .tx_script(send_note_transaction_script)
-        .extend_expected_output_notes(vec![OutputNote::Full(note.clone())])
+        .extend_expected_output_notes(vec![RawOutputNote::Full(note.clone())])
         .build()?
         .execute()
         .await?;
 
-    assert_eq!(executed_transaction.output_notes().get_note(0), &OutputNote::Full(note));
+    assert_eq!(executed_transaction.output_notes().get_note(0), &RawOutputNote::Full(note));
 
     Ok(())
 }